pyproject.toml
[project]
name = "balthazar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiodns>=3.6.1",
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-genai>=1.56.0",
    "html-to-markdown>=2.16.1",
    "mastodon-py>=2.1.4",
    "orjson>=3.11.5",
    "playwright>=1.57.0",
    "pypdf>=6.5.0",
]
